pyproject.toml
[project]
name = "vaglito-store"
version = "0.1.0"
description = "The API store in FastAPI"
authors = [
    {name = "Cesar Gonzalez",email = "cesargonzalez390@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"
pydantic = "^2.11.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


